process_data/chat_parser/video-to-chatml/requirements.txt
openai-whisper>=20231117
torch>=1.9.0
ffmpeg-python>=0.2.0